feat(CModal): Improve close behavior
Previously closed when mouse up event happened on backdrop. Can lead to a strange behaviour. E.g. user wants to select and copy all text in modal. Selecting all text and mouse up happens outside of the modal just closes the modal and makes it impossble to copy the selected text.

Now only closes the modal when the mousedown and mouseup event happens outside of the modal.